requirements.txt
fonttools>=4.60.0
pillow>=12.0.0
